Final merge for exam

+++ controllers/intermittentFasting.js
@@ -1,10 +1,14 @@
-const { intermittentFastingModel } = require("../models");
+const { intermittentFastingModel, usersModel } = require("../models");
 const { handleHttpError } = require("../utils/handleErrors");
 const schedule = require("node-schedule");
 const { sendIntermittentFastingNotificationEmail } = require("./notifications");
 
 const createIntermittentFasting = async (req, res) => {
   try {
+    const userId = req.userId;
+    const userData = await usersModel.findById(userId);
+    const email = userData.email;
+    const userName = userData.firstName;
     const overlappingFasting = await intermittentFastingModel.findOne({
       $or: [
         {
@@ -12,22 +16,22 @@ const createIntermittentFasting = async (req, res) => {
           endDateTime: { $gt: req.body.startDateTime },
         },
       ],
-      userId: req.body.userId,
+      userId: userId,
     });
 
     if (overlappingFasting) {
       return handleHttpError(res, "CONFLICTING_FASTING_PERIOD", 501);
     }
-
-    const data = await intermittentFastingModel.create(req.body);
+    const itFastingData = { ...req.body, userId };
+    const data = await intermittentFastingModel.create(itFastingData);
     const endDateTime = new Date(req.body.endDateTime);
     schedule.scheduleJob(
-      endDateTime.setTime(endDateTime.getTime() - 60 * 60000),
+      endDateTime.setTime(endDateTime.getTime() + 2 * 3600000),
       () => {
         const reqUpdateUser = {
           body: {
-            email: req.body.email,
-            userName: req.body.userName,
+            email: email,
+            userName: userName,
           },
         };
 
@@ -37,7 +41,6 @@ const createIntermittentFasting = async (req, res) => {
             console.log(`Status Code: ${statusCode}`);
           },
         };
-
         sendIntermittentFastingNotificationEmail(reqUpdateUser, resUpdateUser);
       }
     );
@@ -47,28 +50,40 @@ const createIntermittentFasting = async (req, res) => {
   }
 };
 
-const getIntermittentFastingByUserId = async (req, res) => {
+const getActiveIntermittentFastingByUserId = async (req, res) => {
   try {
     const data = await intermittentFastingModel.find({
-      userId: req.params.userId,
+      userId: req.userId,
     });
-    res.send({ data });
+    const today = new Date();
+    today.setSeconds(0);
+    today.setHours(today.getHours() - 3);
+    const filteredData = data.find(
+      (item) => today >= item.startDateTime && today <= item.endDateTime
+    );
+    res.send({ filteredData });
   } catch (e) {
-    handleHttpError(res, "ERROR_GET_INTERMITTENT_FASTING_BY_USER_ID", 500);
+    handleHttpError(
+      res,
+      "ERROR_GET_ACTIVE_INTERMITTENT_FASTING_BY_USER_ID",
+      500
+    );
   }
 };
 
-const getActiveIntermittentFastingByUserId = async (req, res) => {
+const getNextIntermittentFastingByUserId = async (req, res) => {
   try {
-    const currentDate = new Date();
+    const today = new Date();
+    today.setSeconds(0);
+    today.setHours(today.getHours() - 3);
+    const filteredData = await intermittentFastingModel
+      .find({
+        userId: req.userId,
+        startDateTime: { $gt: today }, // Filtrar por startDateTime mayor que la fecha actual
+      })
+      .sort({ startDateTime: 1 }) // Ordenar por startDateTime ascendente
+      .limit(1); // Limitar el resultado a 1 documento
 
-    const data = await intermittentFastingModel.find({
-      userId: req.params.userId,
-    });
-    const filteredData = data.filter(
-      (item) =>
-        new Date() >= item.startDateTime && new Date() <= item.endDateTime
-    );
     res.send({ filteredData });
   } catch (e) {
     handleHttpError(
@@ -81,6 +96,15 @@ const getActiveIntermittentFastingByUserId = async (req, res) => {
 
 const deleteActiveIntermittentFasting = async (req, res) => {
   try {
+    userId = req.userId;
+    const intFastToDelete = await intermittentFastingModel.findById({
+      _id: req.params.IntermittentFastingId,
+    });
+    if (!intFastToDelete || intFastToDelete.userId.toString() !== userId) {
+      return res.status(403).json({
+        message: "You don't have permission to cancel this intermitent fasting",
+      });
+    }
     const data = await intermittentFastingModel.delete({
       _id: req.params.IntermittentFastingId,
     });
@@ -92,7 +116,7 @@ const deleteActiveIntermittentFasting = async (req, res) => {
 
 module.exports = {
   createIntermittentFasting,
-  getIntermittentFastingByUserId,
   getActiveIntermittentFastingByUserId,
   deleteActiveIntermittentFasting,
+  getNextIntermittentFastingByUserId,
 };
